requirements.txt
Django==4.0.4
flake8==4.0.1
isort==5.10.1
pre-commit==2.19.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
django-split-settings==1.1.0
python-decouple==3.6
django-debug-toolbar==3.4.0
